+++ back-end/src/entities/chapter-view-stats-daily.entity.ts
@@ -4,13 +4,13 @@ import { Chapters } from './chapters.entity';
 @Entity('chapter_view_stats_daily')
 @Index(['chapter_id', 'bucket_date'], { unique: true })
 export class ChapterViewStatsDaily {
-  @Column({ type: 'int', primary: true })
+  @Column({ type: 'bigint', primary: true })
   chapter_id: number;
 
   @Column({ type: 'timestamp', primary: true })
   bucket_date: Date;
 
-  @Column({ type: 'int', nullable: false, default: 0 })
+  @Column({ type: 'bigint', nullable: false, default: 0 })
   count: number;
 
   @ManyToOne(() => Chapters, chapter => chapter.viewStats)
